pyproject.toml
[tool.poetry]
name = "cbr-bot"
version = "1.0.0"
description = "Test task for Junior Python Developer"
authors = ["Kirill Volychev <volychevk@gmail.com>"]
readme = "README.md"
packages = [{include = "cbr_bot"}]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"
redis = "^5.0.7"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
flake8-black = "^0.3.6"
flake8-commas = "^2.1.0"
flake8-builtins = "^2.1.0"
flake8-variables-names = "^0.0.6"
pep8-naming = "^0.13.3"
reorder-python-imports = "^3.10.0"
autoflake = "^2.2.0"
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
